Added custom error page

+++ src/main/java/com/spmovy/servlet/DeleteReview.java
@@ -1,22 +1,34 @@
 package com.spmovy.servlet;
 
 import com.spmovy.DatabaseUtils;
+import com.spmovy.Utils;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.sql.SQLException;
 
 @WebServlet("/backend/admin/DeleteReview")
 public class DeleteReview extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        DatabaseUtils db = new DatabaseUtils();
         int movieid = Integer.parseInt(request.getParameter("movieid"));
         int reviewid = Integer.parseInt(request.getParameter("reviewid"));
-        db.executeUpdate("delete from reviews where movieID=? and reviewID=?", movieid, reviewid);
-        db.closeConnection();
+        DatabaseUtils db = Utils.getDatabaseUtils(response);
+        if (db == null) return;
+        try {
+            db.executeUpdate("delete from reviews where movieID=? and reviewID=?", movieid, reviewid);
+        } catch (SQLException e) {
+            e.printStackTrace();
+            response.sendRedirect("/error.html");
+            return;
+        } finally {
+            db.closeConnection();
+        }
+
+
         response.sendRedirect(request.getHeader("referer"));
     }
 }
